Create README.md
L form element with the id consoleForm is submitted,
If the values of the HTML input element with the id requestUrl is empty the HTML paragraph element with the id requestUrlErrMsg should have Required* error message. Otherwise, it should have empty.
Set the value entered in the HTML textarea element with the id requestBody as the value for HTTP request body.
If the value of the requestMethod element is POST,
Make HTTP Request (POST method) using fetch using the value of the HTML input element as URL.
If the value of the requestMethod element is PUT,
Make HTTP Request (PUT method) using fetch using the value of the HTML input element as URL.
Set response status (status code) in the HTML input element with the id responseStatus.
Set HTTP response body in the HTML textarea element with the id responseBody as shown in the image.